requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
decorator==4.4.2
Deprecated==1.2.14
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
frozenlist==1.4.0
gunicorn==21.2.0
h5py==3.9.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
limits==3.6.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy<=1.26.0
openai==0.28.1
ordered-set==4.1.0
packaging==23.2
Pillow==10.0.1
proglog==0.1.10
pydub==0.25.1
Pygments==2.16.1
pymongo==3.11.0
python-dotenv==1.0.0
python-http-client==3.3.7
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
scipy<=1.11.3
sendgrid==6.10.0
six==1.16.0
starkbank-ecdsa==2.2.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==3.0.0
wrapt==1.15.0
yarl==1.9.2
